requirements.txt
pandas
psycopg2-binary
PyPDF2
tabula-py
SQLAlchemy
openpyxl
pdfplumber